chore: archive existing verification implementation
