+++ dodepus-casino/local-sim/admin/logs/index.js
@@ -13,7 +13,6 @@ import {
   appendTransactionLog,
   readTransactionLogs,
 } from './transactions.js';
-import { VERIFICATION_LOGS } from './verification.js';
 
 const ADMIN_LOG_SECTIONS = Object.freeze([
   { value: 'overview', label: 'Обзор' },
@@ -24,7 +23,6 @@ const ADMIN_LOG_SECTIONS = Object.freeze([
   { value: 'roles', label: 'Выдать роль' },
   { value: 'role-edit', label: 'Изменить роль' },
   { value: 'transactions', label: 'Транзакции' },
-  { value: 'verification', label: 'Верификация' },
   { value: 'moderators-chat', label: 'Модератор Чат' },
   { value: 'administrators-chat', label: 'Админ Чат' },
   { value: 'staff-chat', label: 'Стаф Чат' },
@@ -61,7 +59,6 @@ const STATIC_SECTION_LOGS = Object.freeze([
   ...PROMO_ARCHIVE_LOGS,
   ...ROLES_LOGS,
   ...ROLE_EDIT_LOGS,
-  ...VERIFICATION_LOGS,
   ...MODERATORS_CHAT_LOGS,
   ...ADMINISTRATORS_CHAT_LOGS,
   ...STAFF_CHAT_LOGS,
