+++ dodepus-casino/src/app/routes.jsx
@@ -13,7 +13,6 @@ import AdminPromoArchive from '../pages/Admin/PromoArchive/index.js';
 import AdminRoles from '../pages/Admin/roles/index.js';
 import AdminRoleEdit from '../pages/Admin/RoleEdit/index.js';
 import AdminTransactions from '../pages/Admin/transactions/index.js';
-import AdminVerification from '../pages/Admin/verification/index.js';
 import AdminModeratorsChat from '../pages/Admin/ModeratorsChat/index.js';
 import AdminAdministratorsChat from '../pages/Admin/AdministratorsChat/index.js';
 import AdminStaffChat from '../pages/Admin/StaffChat/index.js';
@@ -31,7 +30,6 @@ import Personal from '../pages/profile/personal';
 import Wallet from '../pages/profile/wallet';
 import Terminal from '../pages/profile/terminal';
 import History from '../pages/profile/history';
-import Verification from '../pages/profile/verification';
 import Promos from '../pages/profile/promos';
 import Season from '../pages/profile/season';
 import GamesHistory from '../pages/profile/games-history';
@@ -97,7 +95,6 @@ export default function AppRoutes() {
         <Route path="roles" element={<AdminRoles />} />
         <Route path="role-edit" element={<AdminRoleEdit />} />
         <Route path="transactions" element={<AdminTransactions />} />
-        <Route path="verification" element={<AdminVerification />} />
         <Route path="moderators-chat" element={<AdminModeratorsChat />} />
         <Route path="administrators-chat" element={<AdminAdministratorsChat />} />
         <Route path="staff-chat" element={<AdminStaffChat />} />
@@ -129,7 +126,6 @@ export default function AppRoutes() {
         <Route path="history" element={<History />} />
         <Route path="terminal" element={<Terminal />} />
         <Route path="personal" element={<Personal />} />
-        <Route path="verification" element={<Verification />} />
         <Route path="promos" element={<Promos />} />
         <Route path="season" element={<Season />} />
         <Route path="games-history" element={<GamesHistory />} />
