+++ dodepus-casino/src/pages/profile/personal/blocks/AddressBlock/AddressBlock.jsx
@@ -1,7 +1,6 @@
 import { useEffect, useState } from 'react';
 import { Card, Form, Row, Col } from 'react-bootstrap';
 import { useAuth } from '../../../../../app/AuthContext.jsx';
-import { useVerificationModules } from '../../../../../shared/verification/index.js';
 
 const countries = [
   'Украина','Казахстан','Россия','Беларусь',
@@ -13,10 +12,7 @@ const t = (v) => (v ?? '').trim();
 
 export default function AddressBlock() {
   const { user, updateProfile } = useAuth();
-  const { modules: verificationModules = {} } = useVerificationModules(user);
-
-  const addressStatus = String(verificationModules?.address?.status || '').toLowerCase();
-  const addressLocked = addressStatus === 'pending' || addressStatus === 'approved';
+  const addressLocked = false;
 
   const [country, setCountry] = useState(user?.country ?? '');
   const [city, setCity]       = useState(user?.city ?? '');
@@ -92,21 +88,6 @@ export default function AddressBlock() {
               />
             </Col>
           </Row>
-          {addressStatus === 'pending' && (
-            <Form.Text className="text-muted d-block mt-2">
-              Данные отправлены на проверку. Изменить адрес можно после решения администратора.
-            </Form.Text>
-          )}
-          {addressStatus === 'approved' && (
-            <Form.Text className="text-muted d-block mt-2">
-              Адрес подтверждён и заблокирован для редактирования.
-            </Form.Text>
-          )}
-          {addressStatus === 'rejected' && (
-            <Form.Text className="text-warning d-block mt-2">
-              Адрес отклонён. Обновите данные и отправьте их повторно.
-            </Form.Text>
-          )}
         </Form>
       </Card.Body>
     </Card>
