+++ dodepus-casino/src/pages/profile/personal/blocks/GenderDobBlock/GenderDobBlock.jsx
@@ -1,7 +1,6 @@
 import { useMemo, useState, useEffect } from 'react';
 import { Card, Form, Row, Col } from 'react-bootstrap';
 import { useAuth } from '../../../../../app/AuthContext.jsx';
-import { useVerificationModules } from '../../../../../shared/verification/index.js';
 
 const monthNames = ['01','02','03','04','05','06','07','08','09','10','11','12'];
 
@@ -17,10 +16,7 @@ function daysInMonth(m, y) {
 
 export default function GenderDobBlock() {
   const { user, updateProfile } = useAuth();
-  const { modules: verificationModules = {} } = useVerificationModules(user);
-
-  const docStatus = String(verificationModules?.doc?.status || '').toLowerCase();
-  const docLocked = docStatus === 'pending' || docStatus === 'approved';
+  const fieldsLocked = false;
 
   const init = useMemo(() => parseDob(user?.dob), [user?.dob]);
   const normalizeGender = (g) => (g === 'male' || g === 'female' ? g : '');
@@ -55,24 +51,24 @@ export default function GenderDobBlock() {
   useEffect(() => {
     const id = 'block:genderdob';
     window.dispatchEvent(
-      new CustomEvent('personal:dirty', { detail: { id, dirty: !docLocked && !!changed } }),
+      new CustomEvent('personal:dirty', { detail: { id, dirty: !!changed } }),
     );
     return () =>
       window.dispatchEvent(
         new CustomEvent('personal:dirty', { detail: { id, dirty: false } }),
       );
-  }, [changed, docLocked]);
+  }, [changed]);
 
   useEffect(() => {
     const onSave = () => {
-      if (!changed || docLocked) return;
+      if (!changed) return;
       const patch = { dob: newDob };
       if (gender === 'male' || gender === 'female') patch.gender = gender;
       updateProfile(patch);
     };
     window.addEventListener('personal:save', onSave);
     return () => window.removeEventListener('personal:save', onSave);
-  }, [changed, docLocked, gender, newDob, updateProfile]);
+  }, [changed, gender, newDob, updateProfile]);
 
   return (
     <Card>
@@ -86,12 +82,12 @@ export default function GenderDobBlock() {
                 <Form.Check
                   type="radio" id="gender-male" name="gender" label="Мужчина"
                   checked={gender === 'male'} onChange={() => setGender('male')}
-                  disabled={docLocked}
+                  disabled={fieldsLocked}
                 />
                 <Form.Check
                   type="radio" id="gender-female" name="gender" label="Женщина"
                   checked={gender === 'female'} onChange={() => setGender('female')}
-                  disabled={docLocked}
+                  disabled={fieldsLocked}
                 />
               </div>
             </Col>
@@ -99,19 +95,19 @@ export default function GenderDobBlock() {
               <Form.Label>Дата рождения</Form.Label>
               <Row className="g-2">
                 <Col xs={4}>
-                  <Form.Select value={d} onChange={(e) => setD(e.target.value)} disabled={docLocked}>
+                  <Form.Select value={d} onChange={(e) => setD(e.target.value)} disabled={fieldsLocked}>
                     <option value="">ДД</option>
                     {days.map((dd) => <option key={dd} value={dd}>{dd}</option>)}
                   </Form.Select>
                 </Col>
                 <Col xs={4}>
-                  <Form.Select value={m} onChange={(e) => setM(e.target.value)} disabled={docLocked}>
+                  <Form.Select value={m} onChange={(e) => setM(e.target.value)} disabled={fieldsLocked}>
                     <option value="">MM</option>
                     {monthNames.map((mm) => <option key={mm} value={mm}>{mm}</option>)}
                   </Form.Select>
                 </Col>
                 <Col xs={4}>
-                  <Form.Select value={y} onChange={(e) => setY(e.target.value)} disabled={docLocked}>
+                  <Form.Select value={y} onChange={(e) => setY(e.target.value)} disabled={fieldsLocked}>
                     <option value="">ГГГГ</option>
                     {years.map((yy) => <option key={yy} value={yy}>{yy}</option>)}
                   </Form.Select>
@@ -119,21 +115,6 @@ export default function GenderDobBlock() {
               </Row>
             </Col>
           </Row>
-          {docStatus === 'pending' && (
-            <Form.Text className="text-muted d-block mt-2">
-              Данные отправлены на проверку. Изменить дату рождения и пол можно после решения администратора.
-            </Form.Text>
-          )}
-          {docStatus === 'approved' && (
-            <Form.Text className="text-muted d-block mt-2">
-              Документы подтверждены и заблокированы для редактирования.
-            </Form.Text>
-          )}
-          {docStatus === 'rejected' && (
-            <Form.Text className="text-warning d-block mt-2">
-              Документы отклонены. Проверьте данные и отправьте их повторно.
-            </Form.Text>
-          )}
         </Form>
       </Card.Body>
     </Card>
