+++ dodepus-casino/src/pages/profile/personal/blocks/NameBlock/NameBlock.jsx
@@ -1,14 +1,10 @@
 import { useState, useEffect } from 'react';
 import { Card, Form, Row, Col } from 'react-bootstrap';
 import { useAuth } from '../../../../../app/AuthContext.jsx';
-import { useVerificationModules } from '../../../../../shared/verification/index.js';
 
 export default function NameBlock() {
   const { user, updateProfile } = useAuth();
-  const { modules: verificationModules = {} } = useVerificationModules(user);
-
-  const docStatus = String(verificationModules?.doc?.status || '').toLowerCase();
-  const docLocked = docStatus === 'pending' || docStatus === 'approved';
+  const fieldsLocked = false;
 
   const [firstName, setFirstName] = useState(user?.firstName ?? '');
   const [lastName, setLastName] = useState(user?.lastName ?? '');
@@ -23,17 +19,17 @@ export default function NameBlock() {
   useEffect(() => {
     const id = 'block:name';
     window.dispatchEvent(
-      new CustomEvent('personal:dirty', { detail: { id, dirty: !docLocked && !!changed } }),
+      new CustomEvent('personal:dirty', { detail: { id, dirty: !!changed } }),
     );
     return () =>
       window.dispatchEvent(
         new CustomEvent('personal:dirty', { detail: { id, dirty: false } }),
       );
-  }, [changed, docLocked]);
+  }, [changed]);
 
   useEffect(() => {
     const onSave = () => {
-      if (!changed || docLocked) return;
+      if (!changed) return;
       updateProfile({
         firstName: (firstName || '').trim(),
         lastName: (lastName || '').trim(),
@@ -41,7 +37,7 @@ export default function NameBlock() {
     };
     window.addEventListener('personal:save', onSave);
     return () => window.removeEventListener('personal:save', onSave);
-  }, [changed, docLocked, firstName, lastName, updateProfile]);
+  }, [changed, firstName, lastName, updateProfile]);
 
   return (
     <Card>
@@ -56,7 +52,7 @@ export default function NameBlock() {
                 placeholder="Иван"
                 value={firstName}
                 onChange={(e) => setFirstName(e.target.value)}
-                disabled={docLocked}
+                disabled={fieldsLocked}
               />
             </Col>
             <Col md={6}>
@@ -66,25 +62,10 @@ export default function NameBlock() {
                 placeholder="Иванов"
                 value={lastName}
                 onChange={(e) => setLastName(e.target.value)}
-                disabled={docLocked}
+                disabled={fieldsLocked}
               />
             </Col>
           </Row>
-          {docStatus === 'pending' && (
-            <Form.Text className="text-muted d-block mt-2">
-              Данные отправлены на проверку. Изменить ФИО можно после решения администратора.
-            </Form.Text>
-          )}
-          {docStatus === 'approved' && (
-            <Form.Text className="text-muted d-block mt-2">
-              Документы подтверждены и заблокированы для редактирования.
-            </Form.Text>
-          )}
-          {docStatus === 'rejected' && (
-            <Form.Text className="text-warning d-block mt-2">
-              Документы отклонены. Исправьте данные и отправьте их повторно.
-            </Form.Text>
-          )}
         </Form>
       </Card.Body>
     </Card>
