+++ dodepus-casino/verifffff/README.md
@@ -1,0 +1,7 @@
+# Архив верификационных модулей
+
+В этой директории собраны все исходные файлы, связанные с клиентской и админской
+верификацией, включая локальный симулятор и документацию. Логика временно
+отключена из приложения, чтобы реализовать её заново с нуля. При необходимости
+можно восстановить исходный код из соответствующих подпапок (`src/`,
+`local-sim/`, `docs/`).
